+++ Course/Course/Presentation/Container/CourseContainerView.swift
@@ -11,127 +11,6 @@ import Discussion
 import Swinject
 import Theme
 
-struct UpgradeCourseViewMessage: View {
-    let message: String
-    let icon: Image
-    @Binding var coordinate: CGFloat
-    @Binding var collapsed: Bool
-    @Binding var shouldShowUpgradeButton: Bool
-    @Binding var shouldHideMenuBar: Bool
-    let backAction: (() -> Void)?
-    
-    var body: some View {
-        ZStack(alignment: .top) {
-            DynamicOffsetView(
-                coordinate: $coordinate,
-                collapsed: $collapsed,
-                shouldShowUpgradeButton: $shouldShowUpgradeButton,
-                shouldHideMenuBar: $shouldHideMenuBar
-            )
-            ZStack {
-                VStack(spacing: 24) {
-                    icon
-                        .resizable()
-                        .frame(width: 96, height: 96)
-                    Text(message)
-                        .multilineTextAlignment(.center)
-                        .font(Theme.Fonts.bodyLarge)
-                        .foregroundColor(Theme.Colors.textPrimary)
-                }
-                VStack {
-                    Spacer()
-                    StyledButton("Back") {
-                        backAction?()
-                    }
-                }
-            }
-            .padding(24)
-        }
-    }
-}
-
-struct UpgradeCourseView: View {
-    let type: CourseAccessErrorHelperType
-    @Binding private var coordinate: CGFloat
-    @Binding private var collapsed: Bool
-    @Binding private var shouldShowUpgradeButton: Bool
-    @Binding private var shouldHideMenuBar: Bool
-    private var backAction: (() -> Void)?
-
-    init(
-        type: CourseAccessErrorHelperType,
-        coordinate: Binding<CGFloat>,
-        collapsed: Binding<Bool>,
-        shouldShowUpgradeButton: Binding<Bool>,
-        shouldHideMenuBar: Binding<Bool>,
-        backAction: (() -> Void)?
-    ) {
-        self.type = type
-        self._coordinate = coordinate
-        self._collapsed = collapsed
-        self._shouldShowUpgradeButton = shouldShowUpgradeButton
-        self._shouldHideMenuBar = shouldHideMenuBar
-        self.backAction = backAction
-    }
-
-    var body: some View {
-        switch type {
-        case let .upgradeable(date, sku, courseID, pacing, screen), let .auditExpired(date, sku, courseID, pacing, screen):
-            VStack {
-                let message = "Your free audit access to this course expired on \(date?.dateToString(style: .monthDayYear) ?? ""). Please upgrade to continue learning and receive a verified certificate."
-                UpgradeInfoView(
-                    isFindCourseButtonVisible: true,
-                    viewModel: Container.shared.resolve(
-                        UpgradeInfoViewModel.self,
-                        arguments: "", message, sku, courseID, screen, pacing
-                    )!,
-                    headerView: {
-                        VStack(spacing: 20) {
-                            DynamicOffsetView(
-                                coordinate: $coordinate,
-                                collapsed: $collapsed,
-                                shouldShowUpgradeButton: $shouldShowUpgradeButton,
-                                shouldHideMenuBar: $shouldHideMenuBar
-                            )
-                            
-                            VStack {
-                                CoreAssets.upgradeArrowImage.swiftUIImage
-                                    .resizable()
-                                    .frame(width: 96, height: 96)
-                                    .padding(.bottom, 4)
-                            }
-                            .frame(maxWidth: .infinity)
-                        }
-                        .frame(maxWidth: .infinity)
-                    }
-                )
-            }
-        case .startDateError(let date):
-            let message = "This course will begin on \(date?.dateToString(style: .monthDayYear) ?? ""). Come back then to start learning!"
-            UpgradeCourseViewMessage(
-                message: message,
-                icon: CoreAssets.upgradeCalendarImage.swiftUIImage,
-                coordinate: $coordinate,
-                collapsed: $collapsed,
-                shouldShowUpgradeButton: $shouldShowUpgradeButton,
-                shouldHideMenuBar: $shouldHideMenuBar,
-                backAction: backAction
-            )
-        case .isEndDateOld(let date):
-            let message = "Your free audit access to this course expired on \(date.dateToString(style: .monthDayYear))."
-            UpgradeCourseViewMessage(
-                message: message,
-                icon: CoreAssets.upgradeArrowImage.swiftUIImage,
-                coordinate: $coordinate,
-                collapsed: $collapsed,
-                shouldShowUpgradeButton: $shouldShowUpgradeButton,
-                shouldHideMenuBar: $shouldHideMenuBar,
-                backAction: backAction
-            )
-        }
-    }
-}
-
 public struct CourseContainerView: View {
     @ObservedObject
     public var viewModel: CourseContainerViewModel
